backups/simple_cleanup_backup/requirements.txt
# Core Bot Dependencies
python-telegram-bot==20.6
python-dotenv==1.0.0
aiohttp==3.9.1

# Database
asyncpg==0.29.0

# Cryptocurrency and Blockchain
requests==2.31.0
cryptography==41.0.7

# Optional Dependencies
redis==5.0.1
psutil==5.9.6

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1 